Split project up into binary and library
Before this change, a user had to download the source code of the project
and run `make build` in order to get a WASM module.
Now, this is not needed anymore as the project was refactored into a binary
and library part. The commandline binary is a high-level driver for the
creation of the WASM module, which downloads library crate from crates.io during
compilation and creates the index from a JSON file, similar to lunr.js and
elasticlunr.